+++ Examples/Examples/Examples/StackedBarsExample.swift
@@ -1,0 +1,172 @@
+//
+//  StackedBarsExample.swift
+//  Examples
+//
+//  Created by ischuetz on 15/05/15.
+//  Copyright (c) 2015 ivanschuetz. All rights reserved.
+//
+
+import UIKit
+
+class StackedBarsExample: UIViewController {
+
+    private var chart: Chart? // arc
+    
+    let sideSelectorHeight: CGFloat = 50
+
+    private func chart(#horizontal: Bool) -> Chart {
+        let labelSettings = ChartLabelSettings(font: ExamplesDefaults.labelFont)
+        
+        func randomColor() -> UIColor {
+            let r = {CGFloat(Float(arc4random()) / Float(UINT32_MAX))}
+            return UIColor(red: r(), green: r(), blue: r(), alpha: 0.7)
+        }
+        
+        let color0 = randomColor()
+        let color1 = randomColor()
+        let color2 = randomColor()
+        let color3 = randomColor()
+        
+        let barModels = [
+            ChartStackedBarModel(constantAxisValue: ChartAxisValueString("A", order: 1, labelSettings: labelSettings), items: [
+                ChartStackedBarItemModel(title: "Pears", quantity: 20, bgColor: color0),
+                ChartStackedBarItemModel(title: "Apples", quantity: 60, bgColor: color1),
+                ChartStackedBarItemModel(title: "Bananas", quantity: 30, bgColor: color2),
+                ChartStackedBarItemModel(title: "Blueberries", quantity: 20, bgColor: color3)
+                ]),
+            ChartStackedBarModel(constantAxisValue: ChartAxisValueString("B", order: 2, labelSettings: labelSettings), items: [
+                ChartStackedBarItemModel(title: "Pears", quantity: 40, bgColor: color0),
+                ChartStackedBarItemModel(title: "Apples", quantity: 30, bgColor: color1),
+                ChartStackedBarItemModel(title: "Bananas", quantity: 10, bgColor: color2),
+                ChartStackedBarItemModel(title: "Blueberries", quantity: 30, bgColor: color3)
+                ]),
+            ChartStackedBarModel(constantAxisValue: ChartAxisValueString("C", order: 3, labelSettings: labelSettings), items: [
+                ChartStackedBarItemModel(title: "Pears", quantity: 30, bgColor: color0),
+                ChartStackedBarItemModel(title: "Apples", quantity: 50, bgColor: color1),
+                ChartStackedBarItemModel(title: "Bananas", quantity: 20, bgColor: color2),
+                ChartStackedBarItemModel(title: "Blueberries", quantity: 10, bgColor: color3)
+                ]),
+            ChartStackedBarModel(constantAxisValue: ChartAxisValueString("D", order: 4, labelSettings: labelSettings), items: [
+                ChartStackedBarItemModel(title: "Pears", quantity: 10, bgColor: color0),
+                ChartStackedBarItemModel(title: "Apples", quantity: 30, bgColor: color1),
+                ChartStackedBarItemModel(title: "Bananas", quantity: 50, bgColor: color2),
+                ChartStackedBarItemModel(title: "Blueberries", quantity: 5, bgColor: color3)
+                ])
+        ]
+        
+        let (axisValues1, axisValues2) = (
+            Array(stride(from: 0, through: 150, by: 20)).map {ChartAxisValueFloat($0, labelSettings: labelSettings)},
+            [ChartAxisValueString("", order: 0, labelSettings: labelSettings)] + barModels.map{$0.constantAxisValue} + [ChartAxisValueString("", order: 5, labelSettings: labelSettings)]
+        )
+        let (xValues, yValues) = horizontal ? (axisValues1, axisValues2) : (axisValues2, axisValues1)
+        
+        let xModel = ChartAxisModel(axisValues: xValues, axisTitleLabel: ChartAxisLabel(text: "Axis title", settings: labelSettings))
+        let yModel = ChartAxisModel(axisValues: yValues, axisTitleLabel: ChartAxisLabel(text: "Axis title", settings: labelSettings))
+        
+        let frame = ExamplesDefaults.chartFrame(self.view.bounds)
+        let chartFrame = self.chart?.frame ?? CGRectMake(frame.origin.x, frame.origin.y, frame.size.width, frame.size.height - sideSelectorHeight)
+        let coordsSpace = ChartCoordsSpaceLeftBottomSingleAxis(chartSettings: ExamplesDefaults.chartSettings, chartFrame: chartFrame, xModel: xModel, yModel: yModel)
+        let (xAxis, yAxis, innerFrame) = (coordsSpace.xAxis, coordsSpace.yAxis, coordsSpace.chartInnerFrame)
+        
+        let chartStackedBarsLayer = ChartStackedBarsLayer(xAxis: xAxis, yAxis: yAxis, innerFrame: innerFrame, barModels: barModels, horizontal: horizontal, barWidth: 40)
+        
+        let settings = ChartGuideLinesDottedLayerSettings(linesColor: UIColor.blackColor(), linesWidth: ExamplesDefaults.guidelinesWidth, axis: .XAndY)
+        let guidelinesLayer = ChartGuideLinesDottedLayer(axisX: xAxis, yAxis: yAxis, innerFrame: innerFrame, settings: settings)
+        
+        return Chart(
+            frame: chartFrame,
+            layers: [
+                xAxis,
+                yAxis,
+                guidelinesLayer,
+                chartStackedBarsLayer
+            ]
+        )
+    }
+    
+    private func showChart(#horizontal: Bool) {
+        self.chart?.clearView()
+        
+        let chart = self.chart(horizontal: horizontal)
+        self.view.addSubview(chart.view)
+        self.chart = chart
+    }
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        self.showChart(horizontal: true)
+        if let chart = self.chart {
+            let sideSelector = DirSelector(frame: CGRectMake(0, chart.frame.origin.y + chart.frame.size.height, self.view.frame.size.width, self.sideSelectorHeight), controller: self)
+            self.view.addSubview(sideSelector)
+        }
+    }
+    
+    class DirSelector: UIView {
+        
+        let left: UIButton
+        let right: UIButton
+        
+        weak var controller: StackedBarsExample?
+        
+        private let buttonDirs: [UIButton : Bool]
+        
+        init(frame: CGRect, controller: StackedBarsExample) {
+            
+            self.controller = controller
+            
+            self.left = UIButton()
+            self.left.setTitle("Horizontal", forState: .Normal)
+            self.right = UIButton()
+            self.right.setTitle("Vertical", forState: .Normal)
+            
+            self.buttonDirs = [self.left : true, self.right : false]
+            
+            super.init(frame: frame)
+            
+            self.addSubview(self.left)
+            self.addSubview(self.right)
+            
+            for button in [self.left, self.right] {
+                button.titleLabel?.font = ExamplesDefaults.fontWithSize(14)
+                button.setTitleColor(UIColor.blueColor(), forState: .Normal)
+                button.addTarget(self, action: "buttonTapped:", forControlEvents: .TouchUpInside)
+            }
+        }
+        
+        func buttonTapped(sender: UIButton) {
+            let horizontal = sender == self.left ? true : false
+            controller?.showChart(horizontal: horizontal)
+        }
+        
+        override func didMoveToSuperview() {
+            let views = [self.left, self.right]
+            for v in views {
+                v.setTranslatesAutoresizingMaskIntoConstraints(false)
+            }
+            
+            let namedViews = Array(enumerate(views)).map{index, view in
+                ("v\(index)", view)
+            }
+            
+            let viewsDict = namedViews.reduce(Dictionary<String, UIView>()) {(var u, tuple) in
+                u[tuple.0] = tuple.1
+                return u
+            }
+            
+            let buttonsSpace: CGFloat = Env.iPad ? 20 : 10
+            
+            let hConstraintStr = namedViews.reduce("H:|") {str, tuple in
+                "\(str)-(\(buttonsSpace))-[\(tuple.0)]"
+            }
+            
+            let vConstraits = namedViews.flatMap {NSLayoutConstraint.constraintsWithVisualFormat("V:|[\($0.0)]", options: NSLayoutFormatOptions.allZeros, metrics: nil, views: viewsDict)}
+            
+            self.addConstraints(NSLayoutConstraint.constraintsWithVisualFormat(hConstraintStr, options: NSLayoutFormatOptions.allZeros, metrics: nil, views: viewsDict)
+                + vConstraits)
+        }
+        
+        required init(coder aDecoder: NSCoder) {
+            fatalError("init(coder:) has not been implemented")
+        }
+    }
+}